# Use SIMD in AES IGE