--- a/packages/crypto-node/lib/ige256.cpp
+++ b/packages/crypto-node/lib/ige256.cpp
@@ -9,16 +9,12 @@
 
 void ige256_encrypt(napi_env env, uint8_t* in, uint32_t length, uint8_t* key, uint8_t* iv, uint8_t* out) {
     EVP_CIPHER_CTX *ctx;
-    uint32_t i, j;
-
-    uint8_t* iv1;
-    uint8_t* iv2;
+    uint32_t i;
+    int written;
     uint8_t* block;
 
-    int written;
-
-    iv1 = &iv[0];
-    iv2 = &iv[16];
+    v16qi iv1 = *(v16qi*)&iv[0];
+    v16qi iv2 = *(v16qi*)&iv[16];
 
     if (!(ctx = EVP_CIPHER_CTX_new())) {
         THROW_OPENSSL_ERROR
@@ -31,18 +27,16 @@ void ige256_encrypt(napi_env env, uint8_t* in, uint32_t length, uint8_t* key, ui
     for (i = 0; i < length; i += AES_BLOCK_SIZE) {
         block = &out[i];
 
-        for (j = 0; j < AES_BLOCK_SIZE; ++j)
-            block[j] = in[i + j] ^ iv1[j];
+        v16qi v_in = *(v16qi*)&in[i];
+        *(v16qi*)block = v_in ^ iv1;
 
         if (1 != EVP_EncryptUpdate(ctx, block, &written, block, AES_BLOCK_SIZE)) {
             THROW_OPENSSL_ERROR
         }
 
-        for (j = 0; j < AES_BLOCK_SIZE; ++j)
-            block[j] ^= iv2[j];
-
-        iv1 = block;
-        iv2 = &in[i];
+        iv1 = *(v16qi*)block ^ iv2;
+        *(v16qi*)block = iv1;
+        iv2 = v_in;
     }
 
     EVP_CIPHER_CTX_free(ctx);
@@ -50,16 +44,12 @@ void ige256_encrypt(napi_env env, uint8_t* in, uint32_t length, uint8_t* key, ui
 
 void ige256_decrypt(napi_env env, uint8_t* in, uint32_t length, uint8_t* key, uint8_t* iv, uint8_t* out) {
     EVP_CIPHER_CTX *ctx;
-    uint32_t i, j;
-
-    uint8_t* iv1;
-    uint8_t* iv2;
+    uint32_t i;
+    int written;
     uint8_t* block;
 
-    int written;
-
-    iv1 = &iv[16];
-    iv2 = &iv[0];
+    v16qi iv1 = *(v16qi*)&iv[16];
+    v16qi iv2 = *(v16qi*)&iv[0];
 
     if (!(ctx = EVP_CIPHER_CTX_new())) {
         THROW_OPENSSL_ERROR
@@ -76,17 +66,15 @@ void ige256_decrypt(napi_env env, uint8_t* in, uint32_t length, uint8_t* key, ui
     for (i = 0; i < length; i += AES_BLOCK_SIZE) {
         block = &out[i];
 
-        for (j = 0; j < AES_BLOCK_SIZE; ++j)
-            block[j] = in[i + j] ^ iv1[j];
+        v16qi v_in = *(v16qi*)&in[i];
+        *(v16qi*)block = v_in ^ iv1;
 
         if (1 != EVP_DecryptUpdate(ctx, block, &written, block, AES_BLOCK_SIZE)) {
             THROW_OPENSSL_ERROR
         }
 
-        for (j = 0; j < AES_BLOCK_SIZE; ++j)
-            block[j] ^= iv2[j];
-
-        iv1 = block;
-        iv2 = &in[i];
+        iv1 = *(v16qi*)block ^ iv2;
+        *(v16qi*)block = iv1;
+        iv2 = v_in;
     }
 }