--- a/packages/wasm/lib/crypto/ctr256.c
+++ b/packages/wasm/lib/crypto/ctr256.c
@@ -1,8 +1,8 @@
 #include "aes256.h"
 
 struct ctr256_ctx {
-    uint32_t expandedKey[EXPANDED_KEY_SIZE];
-    uint8_t iv[AES_BLOCK_SIZE];
+    alignas(16) uint32_t expandedKey[EXPANDED_KEY_SIZE];
+    alignas(16) uint8_t iv[AES_BLOCK_SIZE];
     uint8_t state;
 };
 
@@ -21,13 +21,13 @@ WASM_EXPORT void ctr256_free(struct ctr256_ctx* ctx) {
 }
 
 WASM_EXPORT void ctr256(struct ctr256_ctx* ctx, uint8_t* in, uint32_t length, uint8_t *out) {
-    uint8_t chunk[AES_BLOCK_SIZE];
+    alignas(16) uint8_t chunk[AES_BLOCK_SIZE];
     uint32_t* expandedKey = ctx->expandedKey;
     uint8_t* iv = ctx->iv;
     uint8_t state = ctx->state;
     uint32_t i, j, k;
 
-    aes256_encrypt(iv, chunk, expandedKey);
+    *(v16qi*)chunk = aes256_encrypt(*(v16qi*)iv, expandedKey);
 
     for (i = 0; i < length; i += AES_BLOCK_SIZE) {
         for (j = 0; j < MIN(length - i, AES_BLOCK_SIZE); ++j) {
@@ -42,7 +42,7 @@ WASM_EXPORT void ctr256(struct ctr256_ctx* ctx, uint8_t* in, uint32_t length, ui
                     if (++iv[k])
                         break;
 
-                aes256_encrypt(iv, chunk, expandedKey);
+                *(v16qi*)chunk = aes256_encrypt(*(v16qi*)iv, expandedKey);
             }
         }
     }