--- a/packages/wasm/lib/crypto/ige256.c
+++ b/packages/wasm/lib/crypto/ige256.c
@@ -1,59 +1,47 @@
 #include "aes256.h"
 
 WASM_EXPORT void ige256_encrypt(uint8_t* in, uint32_t length, uint8_t* out) {
-    uint32_t expandedKey[EXPANDED_KEY_SIZE];
-    uint32_t i, j;
+    alignas(16) uint32_t expandedKey[EXPANDED_KEY_SIZE];
+    uint32_t i;
 
-    uint8_t* iv1;
-    uint8_t* iv2;
-    uint8_t* block;
-
-    iv1 = &aes_shared_iv_buffer[0];
-    iv2 = &aes_shared_iv_buffer[16];
+    v16qi iv1 = *(v16qi*)&aes_shared_iv_buffer[0];
+    v16qi iv2 = *(v16qi*)&aes_shared_iv_buffer[16];
 
     aes256_set_encryption_key(aes_shared_key_buffer, expandedKey);
 
     for (i = 0; i < length; i += AES_BLOCK_SIZE) {
-        block = &out[i];
+        v16qi v_in = *(v16qi*)&in[i];
 
-        for (j = 0; j < AES_BLOCK_SIZE; ++j)
-            block[j] = in[i + j] ^ iv1[j];
+        v16qi block = aes256_encrypt(v_in ^ iv1, expandedKey);
 
-        aes256_encrypt(block, block, expandedKey);
+        block ^= iv2;
 
-        for (j = 0; j < AES_BLOCK_SIZE; ++j)
-            block[j] ^= iv2[j];
+        *(v16qi*)&out[i] = block;
 
         iv1 = block;
-        iv2 = &in[i];
+        iv2 = v_in;
     }
 }
 
 WASM_EXPORT void ige256_decrypt(uint8_t* in, uint32_t length, uint8_t* out) {
-    uint32_t expandedKey[EXPANDED_KEY_SIZE];
-    uint32_t i, j;
+    alignas(16) uint32_t expandedKey[EXPANDED_KEY_SIZE];
+    uint32_t i;
 
-    uint8_t* iv1;
-    uint8_t* iv2;
-    uint8_t* block;
-
-    iv1 = &aes_shared_iv_buffer[16];
-    iv2 = &aes_shared_iv_buffer[0];
+    v16qi iv1 = *(v16qi*)&aes_shared_iv_buffer[16];
+    v16qi iv2 = *(v16qi*)&aes_shared_iv_buffer[0];
 
     aes256_set_decryption_key(aes_shared_key_buffer, expandedKey);
 
     for (i = 0; i < length; i += AES_BLOCK_SIZE) {
-        block = &out[i];
+        v16qi v_in = *(v16qi*)&in[i];
 
-        for (j = 0; j < AES_BLOCK_SIZE; ++j)
-            block[j] = in[i + j] ^ iv1[j];
+        v16qi block = aes256_decrypt(v_in ^ iv1, expandedKey);
 
-        aes256_decrypt(block, block, expandedKey);
+        block ^= iv2;
 
-        for (j = 0; j < AES_BLOCK_SIZE; ++j)
-            block[j] ^= iv2[j];
+        *(v16qi*)&out[i] = block;
 
         iv1 = block;
-        iv2 = &in[i];
+        iv2 = v_in;
     }
 }