Example of output with clang -O3 for x86-64:

before:

```assembly
.LBB7_2:
        mov     ebp, r13d
        lea     r12, [rbx + rbp]
        lea     rdx, [r14 + rbp]
        mov     qword ptr [rsp + 8], rdx
        xor     cl, byte ptr [r14 + rbp]
        mov     byte ptr [rbx + rbp], cl
        movzx   ecx, byte ptr [rax + 1]
        xor     cl, byte ptr [r14 + rbp + 1]
        mov     byte ptr [rbx + rbp + 1], cl
        movzx   ecx, byte ptr [rax + 2]
        xor     cl, byte ptr [r14 + rbp + 2]
        mov     byte ptr [rbx + rbp + 2], cl
...
        xor     cl, byte ptr [r14 + rbp + 14]
        mov     byte ptr [rbx + rbp + 14], cl
        movzx   eax, byte ptr [rax + 15]
        xor     al, byte ptr [r14 + rbp + 15]
        mov     byte ptr [rbx + rbp + 15], al
        mov     rdi, r12
        mov     rsi, r12
        lea     rdx, [rsp + 16]
        call    aes256_decrypt
        movzx   ecx, byte ptr [rbx + rbp]
        xor     cl, byte ptr [r15]
        mov     byte ptr [rbx + rbp], cl
        movzx   eax, byte ptr [r15 + 1]
        xor     byte ptr [rbx + rbp + 1], al
...
        movzx   eax, byte ptr [r15 + 14]
        xor     byte ptr [rbx + rbp + 14], al
        movzx   eax, byte ptr [r15 + 15]
        xor     byte ptr [rbx + rbp + 15], al
        mov     ebp, dword ptr [rsp + 4]
        add     r13d, 16
        mov     r15, qword ptr [rsp + 8]
        mov     rax, r12
        cmp     r13d, ebp
        jb      .LBB7_2
```

after:

```assembly
.LBB7_2:
        movaps  xmm1, xmmword ptr [r14 + r13]
        movaps  xmmword ptr [rsp], xmm1
        xorps   xmm0, xmm1
        mov     rdi, r15
        call    aes256_decrypt
        xorps   xmm0, xmmword ptr [rsp + 16]
        movaps  xmmword ptr [rbx + r13], xmm0
        add     r13, 16
        movaps  xmm1, xmmword ptr [rsp]
        movaps  xmmword ptr [rsp + 16], xmm1
        cmp     r13, r12
        jb      .LBB7_2
```

## benchmarks

### native node addon

before:

```
Benchmarking 2.medium.json

5. openssl-based ige
ige256_encrypt x 17036.43 ops/sec ± 202.43 (211.822 MB/s)
ige256_decrypt x 17253.59 ops/sec ± 151.67 (214.522 MB/s)
Sampled 8199 chunks (101.942 MB) in 10.009 seconds
Average memory usage: 72.56 MB
```

after:

```
Benchmarking 2.medium.json

5. openssl-based ige
ige256_encrypt x 22265.17 ops/sec ± 401.12 (276.590 MB/s)
ige256_decrypt x 22339.11 ops/sec ± 303.75 (277.509 MB/s)
Sampled 8423 chunks (104.634 MB) in 10.007 seconds
Average memory usage: 73.29 MB
```

### wasm

before:

```
Benchmarking 1.small.json

3. tgcrypto-wasm
ige256_encrypt x 114544.36 ops/sec ± 2290.37 (49.097 MB/s)
ige256_decrypt x 128522.47 ops/sec ± 2147.15 (55.088 MB/s)
Sampled 94280 chunks (40.411 MB) in 10.004 seconds
Average memory usage: 78.99 MB

Benchmarking 3.large.json

3. tgcrypto-wasm
ige256_encrypt x 1122.46 ops/sec ± 16.20 (77.865 MB/s)
ige256_decrypt x 1169.30 ops/sec ± 16.06 (81.114 MB/s)
Sampled 494 chunks (34.269 MB) in 10.105 seconds
Average memory usage: 90.48 MB
```

after:

```
Benchmarking 1.small.json

3. tgcrypto-wasm
ige256_encrypt x 136231.63 ops/sec ± 2251.24 (58.382 MB/s)
ige256_decrypt x 153277.59 ops/sec ± 2018.17 (65.687 MB/s)
Sampled 98391 chunks (42.165 MB) in 10.004 seconds
Average memory usage: 75.36 MB

Benchmarking 3.large.json

3. tgcrypto-wasm
ige256_encrypt x 1479.54 ops/sec ± 31.15 (102.703 MB/s)
ige256_decrypt x 1512.70 ops/sec ± 25.64 (105.005 MB/s)
Sampled 527 chunks (36.582 MB) in 10.084 seconds
Average memory usage: 94.57 MB
```

I didn't touch the AES encryption itself, although this software implementation
is probably not the most optimal even without considering intrinsics (there
were some published SIMD methods). The code of the CTR mode looks to be
especially far from optimal.