@teidesu 2024-06-28
i specifically avoided using simd because wasm simd support is... meh to say the least

i think we should build two wasm blobs, one with simd and one without, check for simd support at runtime and depending on that load different wasm blobs. 

something like this can be used to detect simd (~~stolen~~ taken from GoogleChromeLabs/wasm-feature-detect, Apache2 license):
```
WebAssembly.validate(
  new Uint8Array([
    0, 97, 115, 109, 1, 0, 0, 0, 1, 5, 1, 96, 0, 1, 123, 3, 2, 1, 0, 10, 10,
    1, 8, 0, 65, 0, 253, 15, 253, 98, 11,
  ]),
)
```

regarding the aes impl - it's mostly taken from tgcrypto without much modifications. aes-ni inrinsics in wasm are still just a proposal. ctr impl perf is a known issue, see #44. 

btw - what platform did you compile the wasm blob on? for some reason, xcode clang tends to generate more performant code than clang on other platforms...