@Bannerets 2024-06-28
i thought wasm simd support already be good in 2024 (https://webassembly.org/features/#table-row-simd). hmm, i'm not sure, creating two binaries looks to be a hassle for an insufficiently large gain.

> regarding the aes impl - it's mostly taken from tgcrypto without much modifications. aes-ni inrinsics in wasm are still just a proposal.

yeah, I realized that the implementation is mostly tgcrypto when I cloned the benchmarks repository. I understand that we can't use intrinsics. there are some sophisticated simd software implementations but those are much more difficult. the ctr implementation for one doesn't look to really utilize instruction-level parallelism (ctr is parallelizable)

> btw - what platform did you compile the wasm blob on? for some reason, xcode clang tends to generate more performant code than clang on other platforms...

macOS, with binaryen installed. (the benchmarks are done with all binaries built on the same platform though, the one from the repository wasn't used.)