@teidesu 2024-06-28
> insufficiently large gain 

tbh true. it's unlikely the crypto is the bottleneck right now. 20% gain is definitely cool but im not sure if it's even worth it without properly profiling the library first... 
50-70 mib/s is already quite fast and the bottleneck in file upload/download probably lies somewhere in networking...