@eilvelia 2025-03-25
(merged in https://github.com/mtcute/mtcute/commit/cbf9cb867d99488580ce59943d220d77478191b5)